refactor: moved constants and config code to shared library
