+++ src/SurviveTheHuntClient/PlayerState.cs
@@ -1,10 +1,6 @@
 ﻿using CitizenFX.Core;
-using CitizenFX.Core.Native;
+using SurviveTheHuntShared.Utils;
 using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 using static CitizenFX.Core.Native.API;
